requirements.txt
# Core dependencies for AI Call Secretary
# Voice processing
whisper==1.0.0
onnxruntime==1.15.1
pyttsx3==2.90
pydub==0.25.1
librosa==0.10.1
numpy>=1.20.0
scipy>=1.9.0

# LLM integration
requests>=2.28.0
langchain>=0.0.267
ollama>=0.1.0

# Telephony
# ESL==1.0.0  # FreeSwitch Event Socket Library - Install separately via FreeSWITCH build

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# API and web
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
websockets>=11.0.3

# Testing
pytest>=7.4.0
pytest-mock>=3.11.1